--- main.js
@@ -1,257 +1,0 @@
-
-document.addEventListener('DOMContentLoaded', function () {
-
-    // сохранение ссылки API в переменной
-    const apiUrl = 'https://wedev-api.sky.pro/api/v1/sergei-smirnov/comments';
-
-    // получение ссылок на элементы формы и сообщения о загрузке
-    const commentForm = document.querySelector('.add-form');
-    const addButton = document.querySelector('.add-form-button');
-    const nameInput = document.querySelector('.add-form-name');
-    const textInput = document.querySelector('.add-form-text');
-    const loadingMessage = document.querySelector('.loading-message');
-    const addingMessage = document.querySelector('.adding-message');
-    const commentsList = document.querySelector('.comments');
-
-    // массив для хранения комментариев
-    let comments = [];
-
-    // Функция загрузки комментариев и даты с сервера
-    function fetchComments() {
-        loadingMessage.classList.add('active'); // показываем сообщение о загрузке
-        fetch(apiUrl) // выполняем GET-запрос к API для получения комментариев
-            .then(response => {
-                // проверяем, если сервер вернул статус 500 (внутренняя ошибка сервера)
-                if (response.status === 500) {
-                    //  ошибку с сообщением "Сервер поломался"
-                    throw new Error("Сервер поломался");
-                }
-                // если статус ответа не 500, преобразуем ответ в формат JSON
-                return response.json();
-            })
-            .then(data => {
-                // преобразуем данные комментариев в удобный формат и сохраняем в массив
-                comments = data.comments.map(comment => ({
-                    name: comment.author.name,
-                    date: new Date(comment.date).toLocaleString('ru-RU', {
-                        day: '2-digit',
-                        month: '2-digit',
-                        year: '2-digit',
-                        hour: '2-digit',
-                        minute: '2-digit'
-                    }),
-                    text: comment.text,
-                    likes: comment.likes,
-                    liked: comment.isLiked,
-                    id: comment.id,
-                }));
-                // рендерим комментарии на странице
-                renderComments(comments);
-            })
-            .finally(() => {
-                // убираем сообщение о загрузке после завершения запроса
-                loadingMessage.classList.remove('active');
-            })
-            .catch(error => {
-                // если ошибка связана с сервером, показываем пользователю сообщение об ошибке
-                if (error.message === "Сервер поломался") {
-                    alert("Сервер поломался, попробуйте позже");
-                    loadingMessage.textContent = "Комментарии не загрузились";
-                } else {
-                    alert("Ошибка при загрузке комментариев, попробуйте позже");
-                    console.error(error);
-                }
-            });
-    }
-
-
-    // функция рендеринга комментариев
-    function renderComments(commentsArray) {
-        commentsList.innerHTML = ''; // очищаем список комментариев
-
-        if (!commentsArray) return; // проверяем наличие комментариев
-
-        commentsArray.forEach((comment, index) => {
-            // создаем HTML для каждого комментария и добавляем его в список
-            const li = document.createElement('li');
-            li.classList.add('comment');
-            li.innerHTML = `
-              <div class="comment-header">
-                  <div>${comment.name}</div>
-                  <div>${comment.date}</div>
-              </div>
-              <div class="comment-body">
-                  <div class="comment-text">${comment.text}</div>
-              </div>
-              <div class="comment-footer">
-                  <div class="likes">
-                      <span class="likes-counter">${comment.likes}</span>
-                      <button class="like-button${comment.liked ? ' -active-like' : ''}" data-index="${index}"></button>
-                  </div>
-              </div>
-          `;
-            commentsList.appendChild(li);
-
-
-            // обработчик события для кнопки лайка
-            const likeButton = li.querySelector('.like-button');
-            likeButton.addEventListener('click', () => {
-                toggleLike(index)
-                    .then(() => {
-                        // перерисовываем комментарии с обновленными лайками
-                        renderComments(comments);
-                    });
-            });
-
-            // // если состояние лайка загружается, добавляем анимацию
-            // if (comment.isLikeLoading) {
-            //   likeButton.classList.add('loading-like');
-            // } else {
-            //   likeButton.classList.remove('loading-like');
-            // }
-        });
-    }
-
-
-
-    // функция для переключения лайка локально
-    function toggleLike(index) {
-        const comment = comments[index];
-        comment.isLikeLoading = true; // устанавливаем состояние загрузки
-
-        // const likeButton = document.querySelector(`.like-button[data-index="${index}"]`);
-        // likeButton.classList.add('loading-like'); 
-        // добавляем класс для анимации
-
-        return new Promise(resolve => {
-            setTimeout(() => {
-                // имитируем задержку на сервере и обновляем состояние лайка
-                if (!comment.liked) {
-                    comment.likes++;
-                } else {
-                    comment.likes--;
-                }
-                comment.liked = !comment.liked;
-                comment.isLikeLoading = false; // снимаем состояние загрузки
-
-                // likeButton.classList.remove('loading-like'); // удаляем класс для анимации
-                resolve();
-            }, 2000);
-        });
-    }
-
-
-
-
-    //функция для отправки нового комментария на сервер 
-    function postComment(name, text) {
-        commentForm.classList.add('hidden'); // скрываем форму ввода комментария
-        addingMessage.classList.add('active'); // показываем сообщение о добавлении комментария
-
-        // Отключаем кнопку "Написать"
-        addButton.disabled = true;
-        addButton.textContent = "Отправка...";
-
-        fetch(apiUrl, {
-            method: 'POST',
-            body: JSON.stringify({
-                name,
-                text,
-                forceError: true //форсеррор для 500 ошибки
-            })
-        })
-            .then(response => {
-                if (response.status === 400) {
-                    addButton.textContent = "Написать";
-                    throw new Error("Вы ввели имя короче 3-х символов");
-                }
-                if (response.status === 500) {
-                    throw new Error("Ошибка сервера");
-                }
-                if (response.status === 201) {
-                    return response.json();
-                };
-            })
-            .then(() => fetchComments()) // загружаем обновлённые данные с сервера
-            .then(() => {
-                clearInputFields(); // очищаем поля ввода
-                renderComments(comments); // перерисовываем комментарии с добавленным
-
-                // Включаем кнопку "Написать" после успешной отправки
-                addButton.disabled = false;
-                addButton.textContent = "Написать";
-            })
-            .finally(() => {
-                commentForm.classList.remove('hidden'); // показываем форму ввода комментария
-                addingMessage.classList.remove('active'); // скрываем сообщение о добавлении комментария
-            })
-            .catch(error => {
-                console.error('Ошибка при добавлении комментария:', error);
-
-                // В случае ошибки включаем кнопку "Написать"
-                addButton.disabled = false;
-                addButton.textContent = "Написать";
-                alert(error.message);
-
-            });
-    }
-
-
-    // проверка и замена вредоносного кода в полях ввода
-    function checkSubmit() {
-        const inputValue = nameInput.value
-            .replaceAll('&', '&amp;')
-            .replaceAll('<', '&lt;')
-            .replaceAll('>', '&gt;')
-            .replaceAll('"', '&quot;');
-
-        const areaFormValue = textInput.value
-            .replaceAll('&', '&amp;')
-            .replaceAll('<', '&lt;')
-            .replaceAll('>', '&gt;')
-            .replaceAll('"', '&quot;')
-            .replaceAll('QUOTE_BEGIN', "<div class='quote'>")
-            .replaceAll('QUOTE_END', '</div>')
-            .replaceAll('\n', '<br>');
-
-        // если поля не пустые, то отправляем комментарий
-        if (inputValue !== '' && areaFormValue !== '') {
-            postComment(inputValue, areaFormValue);
-        }
-    }
-
-    // функция для проверки первого и последнего символа в поле ввода
-    function checkFields() {
-        const name = nameInput.value.trim();
-        const text = textInput.value.trim();
-        return name !== '' && text !== '';
-    }
-
-    // функция сброса полей ввода
-    function clearInputFields() {
-        nameInput.value = '';
-        textInput.value = '';
-        updateButtonState(); // обновление состояния кнопки после очистки полей
-    }
-
-    // функция для обновления состояния кнопки отправки
-    function updateButtonState() {
-        addButton.disabled = !checkFields(); // отключаем кнопку отправки, если поля не заполнены
-    }
-
-    // добавляем обработчики событий для ввода имени и текста комментария
-    nameInput.addEventListener('input', updateButtonState);
-    textInput.addEventListener('input', updateButtonState);
-    addButton.addEventListener('click', checkSubmit); //  обработчик клика по кнопке отправки комментария
-    textInput.addEventListener('keyup', (event) => {
-        if (event.key === 'Enter' && event.shiftKey) {
-            checkSubmit(); //  отправка комментария по нажатию Shift+Enter
-            event.preventDefault();
-        }
-    });
-
-    // загружаем комментарии о очищаем поля при загрузке страницы
-    fetchComments();
-    clearInputFields()
-
-});
